+++ react/components/Charts/typings.d.ts
@@ -1,0 +1,9 @@
+type BaseGridProps = Pick<CartesianGridProps, 'vertical' | 'horizontal'>
+type BaseAxisProps = Pick<XAxisProps, 'axisLine' | 'tickLine'>
+type BaseContainerProps = Pick<ResponsiveContainerProps, 'height' | 'width'>
+
+type DefaultProps = {
+    axis: BaseAxisProps,
+    container: BaseContainerProps,
+    grid: BaseGridProps
+}
